+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add Questionnaire repository
 - Bump fragment-ktx to v1.5.5
 - Bump threeten-extra to v1.7.2
+- Add Questionnaire sync resource
 
 ### Internal